Move AsyncRetryOp traits to a separate file.
This is part of #1256.

These traits will be shared by a AsyncRetryOp's counterpart for polling.
